requirements.txt
flask~=3.1.0
flask-restx~=1.3.0
gevent~=24.11.1
google_cloud_secret_manager~=2.23.1
google_crc32c~=1.6.0
gunicorn~=23.0.0
python-dotenv~=1.0.1
requests~=2.32.3
